+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+## [1.5.15]
+- Added global sources functionality
+- Added xgraph source command to cli
+
 ## [1.5.14]
 - Fixed log rollover error
 - Fixed errors with cache.deleteEntity